Айди теперь типа long, а не int

+++ src/main/java/com/kika/smllybot/database/sql/bank/dto/BankAccount.java
@@ -3,6 +3,6 @@ package com.kika.smllybot.database.sql.bank.dto;
 import java.sql.Timestamp;
 
 public record BankAccount(
-        int id, String name,
+        long id, String name,
         int star, long iris,
         long irisCoin, Timestamp lastFarm) {}
